requirements.txt
absl-py==2.1.0
alembic==1.14.0
annotated-types==0.7.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.0
colorlog==6.9.0
connected-components-3d==3.21.0
contourpy==1.1.1
crccheck==1.3.0
cycler==0.12.1
Cython==3.0.11
dbus-python==1.2.16
deep-pipe==0.4.3
docker-pycreds==0.4.0
edt==2.4.1
eval-type-backport==0.2.2
fastremap==1.15.0
filelock==3.16.1
fonttools==4.55.3
fsspec==2025.3.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.37.0
google-auth-oauthlib==1.0.0
greenlet==3.1.1
grpcio==1.68.1
higra==0.6.10
huggingface-hub==0.30.2
idna==3.10
imageio==2.35.1
imops==0.9.1
importlib-metadata==8.5.0
importlib-resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
lazy-loader==0.4
lazycon==0.6.4
loky==3.4.1
Mako==1.3.8
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.7.5
networkx==3.1
nibabel==5.2.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
optuna==4.1.0
packaging==24.2
pandas==2.0.3
pathlib==1.0.1
patsy==1.0.1
pdp==0.3.0
pillow==9.5.0
platformdirs==4.3.6
protobuf==5.29.1
psutil==7.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.1
pydantic==2.10.6
pydantic-core==2.27.2
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytorch-ignite==0.5.1
pytz==2024.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
resource-manager==0.11.4
rsa==4.9
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==2.27.0
setproctitle==1.3.6
SimpleITK==2.4.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.36
statsmodels==0.14.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboard-easy==0.2.3
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.15
torch==1.13.0
torchaudio==0.8.1
torchvision==0.14.0
tqdm==4.67.1
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wandb==0.19.10
werkzeug==3.0.6
zipp==3.20.2
